requirements.txt
telethon
python-dotenv
Flask